オーバーヘッドを削減
