+++ .idea/misc.xml
@@ -3,6 +3,9 @@
   <component name="Black">
     <option name="sdkName" value="Python 3.11 (MMDMotionMaker)" />
   </component>
+  <component name="CMakePythonSetting">
+    <option name="pythonIntegrationState" value="YES" />
+  </component>
   <component name="CMakeWorkspace" PROJECT_DIR="$PROJECT_DIR$/src">
     <contentRoot DIR="$PROJECT_DIR$" />
   </component>
