--- cmake-build-debug/.cmake/api/v1/reply/toolchains-v1-5566bc40437a6707c7eb.json
@@ -1,52 +1,0 @@
-{
-	"kind" : "toolchains",
-	"toolchains" : 
-	[
-		{
-			"compiler" : 
-			{
-				"id" : "GNU",
-				"implicit" : 
-				{
-					"includeDirectories" : 
-					[
-						"/usr/lib/gcc/x86_64-linux-gnu/13/include",
-						"/usr/local/include",
-						"/usr/include/x86_64-linux-gnu",
-						"/usr/include"
-					],
-					"linkDirectories" : 
-					[
-						"/usr/lib/gcc/x86_64-linux-gnu/13",
-						"/usr/lib/x86_64-linux-gnu",
-						"/usr/lib",
-						"/lib/x86_64-linux-gnu",
-						"/lib"
-					],
-					"linkFrameworkDirectories" : [],
-					"linkLibraries" : 
-					[
-						"gcc",
-						"gcc_s",
-						"c",
-						"gcc",
-						"gcc_s"
-					]
-				},
-				"path" : "/usr/bin/cc",
-				"version" : "13.2.0"
-			},
-			"language" : "C",
-			"sourceFileExtensions" : 
-			[
-				"c",
-				"m"
-			]
-		}
-	],
-	"version" : 
-	{
-		"major" : 1,
-		"minor" : 0
-	}
-}
