+++ cmake-build-debug/.ninja_log
@@ -1,13 +1,2 @@
 # ninja log v5
-0	21	1705847185744117329	CMakeFiles/test.dir/main.c.o	43739a7ec2f5c8db
-21	43	1705847185764117709	test	b61fbb9cbdfbd1ce
-0	21	1705847194428281619	CMakeFiles/test.dir/main.c.o	43739a7ec2f5c8db
-21	38	1705847194444281920	test	b61fbb9cbdfbd1ce
-1	26	1705847207572527780	CMakeFiles/test.dir/main.c.o	43739a7ec2f5c8db
-26	45	1705847207592528152	test	b61fbb9cbdfbd1ce
-0	25	1705847213744642347	CMakeFiles/test.dir/main.c.o	43739a7ec2f5c8db
-25	42	1705847213764642718	test	b61fbb9cbdfbd1ce
-0	26	1705847217028703045	CMakeFiles/test.dir/main.c.o	43739a7ec2f5c8db
-26	43	1705847217044703340	test	b61fbb9cbdfbd1ce
-0	22	1705847618687056874	CMakeFiles/test.dir/main.c.o	43739a7ec2f5c8db
 22	38	1705847618703057094	test	b61fbb9cbdfbd1ce
