+++ cmake-build-debug/CMakeFiles/TargetDirectories.txt
@@ -1,3 +1,3 @@
-/home/shuta/IdeaProjects/MMDMotionMaker/cmake-build-debug/CMakeFiles/test.dir
+/home/shuta/IdeaProjects/MMDMotionMaker/cmake-build-debug/CMakeFiles/MMDMotionMaker.dir
 /home/shuta/IdeaProjects/MMDMotionMaker/cmake-build-debug/CMakeFiles/edit_cache.dir
 /home/shuta/IdeaProjects/MMDMotionMaker/cmake-build-debug/CMakeFiles/rebuild_cache.dir
