+++ cmake-build-debug/CMakeFiles/clion-environment.txt
@@ -1,3 +1,4 @@
-ToolSet: 1.0 (local)Options: 
+ToolSet: 1.0 (local)Ninja: 1.11.1@/home/shuta/.local/share/JetBrains/Toolbox/apps/clion/bin/ninja/linux/x64/ninja
+Options: 
 
-Options:
+Options:-DCMAKE_MAKE_PROGRAM=/home/shuta/.local/share/JetBrains/Toolbox/apps/clion/bin/ninja/linux/x64/ninja
