+++ src/CMakeLists.txt
@@ -5,9 +5,11 @@ set(CMAKE_C_STANDARD 11)
 
 add_executable(test
         vmd/vmdLoader.c
-        vmd/vmdStruct.c
+        vmd/vmdStruct.h
         vmd/vmdLoader.c
-        wrapper.c)
+        vmd/indexlib.c
+        pmx/modelLoader.c
+        pmx/pmxStruct.h)
 target_link_libraries(test json-c m)
 include_directories(/usr/include/python3.11, vmd)
