+++ src/setup.py
@@ -4,9 +4,8 @@ setup(
     ext_modules=[
         Extension(
             name="VMDConverter",
-            sources=["wrapper.c", "vmd/vmdLoader.c", "modelLoader.c", "vmd/vmdStruct.c"],
-            include_dirs=["vmd"],
-            libraries=['vmd/indexlib.c', 'json-c'],
+            sources=["wrapper.c", "vmd/vmdLoader.c", "pmx/modelLoader.c"],
+            include_dirs=["vmd", "pmx"],
         ),
     ]
 )
