+++ src/vmd/vmdStruct.h
@@ -1,0 +1,39 @@
+#ifndef vmdStruct_H
+#define vmdStruct_H
+
+#pragma pack(1) // 構造体をきつくパッキングし、1バイトのアライメント
+#pragma once
+
+// ヘッダ
+struct Header_vmd{
+    char header[30]; // "Vocaloid Motion Data 0002\0\0\0\0\0" 30byte
+// (MMDver2以前のvmdは"Vocaloid Motion Data file\0")
+    char modelName[20]; // モデル名 20byte(MMDver2以前のvmdは10byte)
+// 内容がカメラ,照明,セルフ影の場合は"カメラ・照明\0on Data"となる
+};
+
+struct MaxFrame {
+    int maxFrame;
+};
+
+//ボーンキーフレーム要素データ(111Bytes/要素)
+struct BoneFrame{
+    char name[15]; // "センター\0"などのボーン名の文字列 15byte
+    int frame; // フレーム番号
+    float x; // ボーンのX軸位置,位置データがない場合は0
+    float y; // ボーンのY軸位置,位置データがない場合は0
+    float z; // ボーンのZ軸位置,位置データがない場合は0
+    float qx; // ボーンのクォータニオンX回転,データがない場合は0
+    float qy; // ボーンのクォータニオンY回転,データがない場合は0
+    float qz; // ボーンのクォータニオンZ回転,データがない場合は0
+    float qw; // ボーンのクォータニオンW回転,データがない場合は1
+    char bezier[64]; // 補間パラメータ
+};
+
+typedef struct {
+    struct Header_vmd header;
+    struct MaxFrame maxFrame;
+    struct BoneFrame *boneFrame;
+} MotionData;
+
+#endif
